[refactor]: Move Iroha instance functions into `cli` crate.
