+++ cli/Cargo.toml
@@ -16,7 +16,7 @@ bridge = ["iroha_core/bridge"]
 dex = ["iroha_core/dex"]
 roles = ["iroha_core/roles", "iroha_permissions_validators/roles"]
 telemetry = ["iroha_core/telemetry"]
-dev-telemetry = ["iroha_core/dev-telemetry"]
+dev-telemetry = ["iroha_core/dev-telemetry", "iroha_telemetry"]
 default = ["bridge"]
 
 [badges]
@@ -25,9 +25,14 @@ is-it-maintained-open-issues = { repository = "https://github.com/hyperledger/ir
 maintenance = { status = "actively-developed" }
 
 [dependencies]
+iroha_futures = { path = "../futures" }
+iroha_actor = { path = "../actor" }
 iroha_core = { version = "=2.0.0-pre.1", path = "../core" }
+iroha_data_model = { path = "../data_model" }
 iroha_permissions_validators = { version = "=2.0.0-pre.1", path = "../permissions_validators" }
 iroha_logger = { version = "=2.0.0-pre.1", path = "../logger"}
+iroha_telemetry = { version = "=2.0.0-pre.1", path = "../telemetry", optional = true }
 
 color-eyre = "0.5.11"
 tokio = { version = "1.6.0", features = ["rt", "rt-multi-thread"]}
+eyre = "0.6.5"
