+++ client/tests/integration/config.rs
@@ -7,7 +7,7 @@ use test_network::{Peer as TestPeer, *};
 fn get_config() {
     // The underscored variables must not be dropped until end of closure.
     let (_dont_drop, _dont_drop_either, test_client) = <TestPeer>::start_test_with_runtime();
-    wait_for_genesis_committed(vec![test_client.clone()], 0);
+    wait_for_genesis_committed(&vec![test_client.clone()], 0);
 
     let field = test_client.get_config_docs(&["torii"]).unwrap().unwrap();
     assert!(field.contains("IROHA_TORII"));
