+++ client/tests/integration/connected_peers.rs
@@ -25,7 +25,7 @@ fn connected_peers_with_f(faults: u64) {
     let mut status;
 
     let (_rt, network, mut genesis_client) = <Network>::start_test_with_runtime(n_peers as u32, 1);
-    wait_for_genesis_committed(network.clients(), 0);
+    wait_for_genesis_committed(&network.clients(), 0);
     let pipeline_time = Configuration::pipeline_time();
 
     // Confirm all peers connected
