+++ client/tests/integration/events/pipeline.rs
@@ -26,7 +26,7 @@ fn test_with_instruction_and_status(
     should_be_committed: bool,
 ) -> Result<()> {
     let (_rt, network, _) = <Network>::start_test_with_runtime(PEER_COUNT.try_into().unwrap(), 1);
-    wait_for_genesis_committed(network.clients(), 0);
+    wait_for_genesis_committed(&network.clients(), 0);
 
     for submitting_peer in 0..PEER_COUNT {
         let pipeline_time = Configuration::pipeline_time();
