+++ client/tests/integration/multiple_blocks_created.rs
@@ -14,7 +14,7 @@ const N_BLOCKS: usize = 510;
 fn long_multiple_blocks_created() {
     // Given
     let (_rt, network, mut iroha_client) = <Network>::start_test_with_runtime(4, 1);
-    wait_for_genesis_committed(network.clients(), 0);
+    wait_for_genesis_committed(&network.clients(), 0);
     let pipeline_time = Configuration::pipeline_time();
 
     let create_domain = RegisterBox::new(IdentifiableBox::Domain(Domain::test("domain").into()));
