+++ client/tests/integration/unregister_peer.rs
@@ -10,7 +10,7 @@ use test_network::*;
 fn network_stable_after_add_and_after_remove_peer() -> Result<()> {
     // Given a network
     let (rt, network, mut genesis_client, pipeline_time, account_id, asset_definition_id) = init()?;
-    wait_for_genesis_committed(network.clients(), 0);
+    wait_for_genesis_committed(&network.clients(), 0);
 
     // When assets are minted
     mint(
